Replace private barrier from test_locking_queue with utils/event
